+++ src/ui/handlers/CommandHandler.ts
@@ -1,6 +1,5 @@
 import * as vscode from 'vscode';
 import { TutorialController } from '../controllers/TutorialController';
-import { TutorialUriHandler } from './UriHandler';
 
 /**
  * Handles VS Code commands related to Gitorial tutorials.
@@ -45,19 +44,6 @@ export class CommandHandler {
       })
     );
     
-    context.subscriptions.push(
-      vscode.commands.registerCommand('gitorial.debug', () => {
-        tempTestExternalUri(this.tutorialController);
-      })
-    );
-
     console.log('Gitorial commands registered.');
   }
 } 
-
-async function tempTestExternalUri(controller: TutorialController) {
-  const uri = vscode.Uri.parse("cursor://AndrzejSulkowski.gitorial/sync?platform=github&creator=shawntabrizi&repo=rust-state-machine&commitHash=b74e58d9b3165a2e18f11f0fead411a754386c75");
-  // codium://AndrzejSulkowski.gitorial/sync?platform=github&creator=shawntabrizi&repo=rust-state-machine&commitHash=b74e58d9b3165a2e18f11f0fead411a754386c75
-  const uriHandler = new TutorialUriHandler(controller);
-  await uriHandler.handleUri(uri);
-}